requirements.txt
asgiref==3.5.2
crispy-bootstrap5==0.6
django==4.1.2
django-crispy-forms==1.14.0
django-filter==21.1
django-safedelete==1.3.1
# django-simple-history==3.0.0
./utils/django-simple-history
django-sslserver==0.22
djangorestframework==3.13.1
importlib-metadata==4.11.3
Markdown==3.3.6
Pillow==9.2.0
psycopg2==2.9.5
python-dotenv==0.19.2
pytz==2022.1
sqlparse==0.4.2
tzdata==2021.5
zipp==3.7.0
PyPDF2==2.10.9
Unidecode==1.3.4
django-q==1.3.9
django-jsignature==0.11
pdfkit==1.0.0
typing_extensions==4.4.0
pandas==1.5.3
numpy==1.23.2
python-json-logger==2.0.4

